chromatic.Corrector: Do not calculate but fit inverse transformation
Previously, only the correction transformation of channel 1 was fit using
linear regression and the transformation for channel 2 was calculated as
the inverse.

Now get both transformations from linear regression. This means, that
there are now two sets of parameters, which both need to be saved as hdf5,
leading to a (backwards-incompatible) change of the file structure.