gui.locator: Change how supported algorithms are enumerated
Create an `algorithm` module that holds a dict(name, (locate, batch)).
Algorithms are referred to by their names in `main_window` and
`locate_options` instead of their locate and batch functions. The reason
for this is that the `multiprocessing` module cannot pass functions as
arguments to other processes. Instead, only the algorithm name is passed
and looked up in the dict in the subprocess itself.