loc.daostorm_3d: Implement "z" model
This allows for fitting the z position during the fitting process, instead
of afterwards from the sigmas of the peaks.

There is no numba accellerated version yet.

The implementation is not 100% compatible with the Zhuang lab one, since
(AFAICS) there the derivative of the exponential factor is not quite right
(they don't divide by the `d` parameter, see the formula in the documentation of
`loc.z_fit.Parameters.exp_factor_der`).