multicolor: Add `find_closest_pairs` function
This uses a KD-tree to find candidates for pairs and then choses those
with the smallest distance.

The function is also used in `find_colocalizations` and
`merge_channels`, which gets rid of the brute force algorithms (which
were slow and memory intensive) and also makes them more robust (e. g.
in presence of multiple candidates).